+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.3
+
+- modified share facebook story method
+- added copy to clipboard
+- Added share on Instagram story with background.
+- Added share on Twitter
+- Added share with Sms
+- Added default share options
+
 ## 0.0.1
 
 - Added share on Instagram story.